# Gamificator/Teaching-HEIGVD-AMT-2016-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Gamification platform
  description: |
    **API specification of the Gamification platform**

    You can try all HTTP operation described in this Swagger spec.
    
    Find source code of this API [here](https://github.com/sarrab/GamifyApp.git)
host: localhost:8090
basePath: /api
schemes:
  - http
  - https
paths:
  /applications:
    get: 
      description: Retourne la liste de toutes les applications enregistrées
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
      responses:
        200:
          description: OK, Applications summary
          schema:
            type: array
            items:
              $ref: '#/definitions/ApplicationDTO'
    post:
      description: Enregistre une application
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Application to add
          schema:
            $ref: '#/definitions/Registration'
          required: true
      responses:
        201:
          description: OK, Application created
  /applications/{applicationId}:
    get:
      description: Retourne l'application dont le nom est spécifié dans le path
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: applicationId
          description: Id de l'application
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Ok, Application details
          schema:
            $ref: '#/definitions/ApplicationDTO'
    put:
      description: Apporte des modifications à l'application
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: applicationId
          description: Id de l'application
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Modification of the application
          schema:
            $ref: '#/definitions/Registration'
          required: false
      responses:
        200:
          description: OK, Application modified
    delete:
      description: Supprime une application
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: applicationId
          description: Id de l'application
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK, Application deleted
  /badges:
    get: 
      description: Retourne la liste de tous les badges de l'application 
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
      responses:
        200:
          description: OK, Badges summary
          schema:
            type: array
            items:
              $ref: '#/definitions/BadgeDTO'
    post:
      description: Crée un badge dans l'application
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Badge to add
          schema:
            $ref: '#/definitions/BadgeDTO'
          required: true
        - $ref: '#/parameters/TokenHeader'
      responses:
        201:
          description: OK, Badge created
  /badges/{badgeId}:
    get:
      description: Retourne le badge dont l'id est spécifié dans le path
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: badgeId
          description: Badge's id
          required: true
          type: integer
          format: int64   
      responses:
        200:
          description: Ok, Badge details
          schema:
            $ref: '#/definitions/BadgeDTO'
    put:
      description: Apporte des modifications au badge
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: badgeId
          description: Badge's id
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Modification of the badge
          schema:
            $ref: '#/definitions/BadgeDTO'
          required: false
      responses:
        200:
          description: OK, Badge modified
    delete:
      description: Supprime le badge
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: badgeId
          description: Badge's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK, Badge deleted
...
definitions:
...
  ApplicationDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        description: Application's name
      badges:
        type: array
        items:
            type: string
            format: uri
        description: Liste de tous les badges d'une application
  Registration:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        description: Application's name
      badges:
        type: array
        items:
            type: string
            format: uri
        description: Liste de tous les badges d'une application
      username:
         type: string
         description: Description du badge
      password:
        type: string
        format: password
        description: Mot de passe pour enregistrer l'application
  BadgeDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        description: Badge name.
      description:
        type: string
        description: Description du badge
      imageURI:
        type: string
        format: uri
        description: Image du badge
      application:
        type: integer
        format: int64
        description: Identifient de l'application à laquelle appartient le badge
...
parameters:
  TokenHeader:
    description: token that identifies the app sending the request
    name: X-Gamification-Token
    in: header
    required: true
    type: string